WIP working on macros
It seemed to be working for the test program, but that was really intended to test other bit-bashing macros.
For some weird reason an argument given the ANTIC DL macros gets brain damaged arguments and attempting to mask out individual bits just fails.
It did not seem that anything had a typo. 
Reworked, and reworked the temporary variables and introduced compile time.prints and it at least showed when this happens, but no reason why.
I made the complaints go away by putting all argument that needs multiple bits OR'd together inside square brackets.   So, this tells me it is something about parsing an expression.